requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart
scikit-learn
requests
python-dotenv
requests
tensorflow-cpu==2.13.0
